feat: refactor task management components and enhance UI interactions
- Replaced TaskModal with TaskDrawer for improved task management experience.
- Introduced TaskLibraryDrawer for better organization of saved tasks.
- Updated ClearCompletedButton icon for consistency in design.
- Enhanced QuickActions and SectionItem components with GSAP animations for better user feedback.
- Refactored Next7Days and DayColumn components to support sections alongside tasks.

These changes improve the overall user experience and streamline task management functionalities.
