+++ src/components/Buttons/index.ts
@@ -1,1 +1,2 @@
 export { Button } from "./Button";
+export { TaskLibraryButton } from "./TaskLibraryButton";
