+++ src/components/Next7Days/TaskItem.tsx
@@ -193,10 +193,6 @@ export const TaskItem = ({
                   key={goal.id}
                   className="inline-flex items-center px-2 py-1 rounded-full text-xs font-medium bg-pri-pur-100 dark:bg-pri-pur-900/30 text-pri-pur-700 dark:text-pri-pur-300 border border-pri-pur-200 dark:border-pri-pur-700/50"
                 >
-                  <Icon
-                    icon="mingcute:target-2-fill"
-                    className="w-3 h-3 mr-1"
-                  />
                   {goal.title}
                 </span>
               ))}
@@ -230,7 +226,7 @@ export const TaskItem = ({
                 }"`}
               >
                 <Icon
-                  icon="mingcute:classify-add-2-fill"
+                  icon="mingcute:classify-add-2-line"
                   width={24}
                   height={24}
                   className="w-5 h-5"
@@ -257,7 +253,7 @@ export const TaskItem = ({
               aria-label={`Delete task "${task.title}"`}
             >
               <Icon
-                icon="mingcute:delete-2-fill"
+                icon="mingcute:delete-back-line"
                 width={24}
                 height={24}
                 className="w-5 h-5"
