+++ src/pages/Goals.tsx
@@ -463,7 +463,7 @@ export function Goals() {
           <div className="flex items-center justify-between gap-4">
             <div className="flex items-center space-x-3">
               <Icon
-                icon="mingcute:target-fill"
+                icon="mingcute:target-line"
                 className="text-pri-pur-500 w-6 h-6 sm:w-8 sm:h-8"
                 aria-hidden="true"
               />
@@ -537,7 +537,7 @@ export function Goals() {
                         tabIndex={4 + index * 6}
                       >
                         <Icon
-                          icon="mingcute:delete-2-fill"
+                          icon="mingcute:delete-back-line"
                           width={20}
                           height={20}
                           className="w-5 h-5 sm:w-6 sm:h-6"
